Zoom into each parent object
Add a tool to move objects into square crops of an image that only includes one parent object each.

The result of this operation can be used for a dedicated object-in-object detector at much higher resolution.

This step must be performed from the output of a very accurate parent-object-detector. Pparents objects such as boxes are both large and typically extremely uniform, so a parent-object-detector is typically the first step anyway since it is so easy. It is better to not train it to also detect the objects inside them unless the object-in-object detector is meant to be co-located with it.