Document startup before app
When assembly code exists, it needs to contain _start code.
Since _start is the first code that gets executed in an app, it
makes more sense to list it before the remaining app code.

If no assembly code exists, the linker uses a _start comes as default with the linker. It calls main() automatically.

* List assembly code before other sources.
* Doctest normal linking when no assembly code exists.
* Doctest -nostartfiles -no-pie linking when assembly code exists.