BUG: Subproject reviewed rather than main project
A review/audit always had too much of the subproject changes within them. This makes it difficult for the user as well as for GPT to attend to the main project changes. Especially when those subprojects actually are submodules that should have their own review process.

We want to focus on the main project changes: It is enough that subprojects PASS.

Also rename makemake.py build.py since we use it to make files in build/. We may want to make it in there instead of next to the Makefile.

* Document downloading and using the generic Makefile.
* Use -m build instead of makemake.py.
* Do not analyze (now) build.py outside its normalize origin.
* Remove subproject reports and diffs.